+++ src/Pixy2CameraUtils.cpp
@@ -1,0 +1,63 @@
+#include "pixy2/Camera.h"
+
+using namespace lima;
+using namespace lima::pixy2;
+
+//---------------------------
+//- Internal
+//---------------------------
+void Camera::demosaic(uint16_t width, uint16_t height, const uint8_t *bayerImage, uint32_t *image)
+{
+  uint32_t x, y, xx, yy, r, g, b;
+  uint8_t *pixel0, *pixel;
+  
+  for (y=0; y<height; y++)
+  {
+    yy = y;
+    if (yy==0)
+      yy++;
+    else if (yy==height-1)
+      yy--;
+    pixel0 = (uint8_t *)bayerImage + yy*width;
+    for (x=0; x<width; x++, image++)
+    {
+      xx = x;
+      if (xx==0)
+	xx++;
+      else if (xx==width-1)
+	xx--;
+      pixel = pixel0 + xx;
+      if (yy&1)
+      {
+        if (xx&1)
+        {
+          r = *pixel;
+          g = (*(pixel-1)+*(pixel+1)+*(pixel+width)+*(pixel-width))>>2;
+          b = (*(pixel-width-1)+*(pixel-width+1)+*(pixel+width-1)+*(pixel+width+1))>>2;
+        }
+        else
+        {
+          r = (*(pixel-1)+*(pixel+1))>>1;
+          g = *pixel;
+          b = (*(pixel-width)+*(pixel+width))>>1;
+        }
+      }
+      else
+      {
+        if (xx&1)
+        {
+          r = (*(pixel-width)+*(pixel+width))>>1;
+          g = *pixel;
+          b = (*(pixel-1)+*(pixel+1))>>1;
+        }
+        else
+        {
+          r = (*(pixel-width-1)+*(pixel-width+1)+*(pixel+width-1)+*(pixel+width+1))>>2;
+          g = (*(pixel-1)+*(pixel+1)+*(pixel+width)+*(pixel-width))>>2;
+          b = *pixel;
+        }
+      }
+      *image = (b<<16) | (g<<8) | r; 
+    }
+  }
+}
